InpaintingModule/requirements.txt
torch
torchvision
torch-fidelity
torchsummary
matplotlib
numpy
pandas
pillow
tqdm
python-telegram-bot
transformers
timm
onnx
onnxruntime
streamlit